Update camera and added movement
Big update, like i saied in the #5 issue, i've fix it with the update of the basic.vert.glsl file!

i've update the transform class to allow movement and added 3d movement to the camera.

windows class can now ouput a delta time, wich is the time between each frame!