FPS camera fix and is now usable
FINALLY, we have fix #6 and now we can start working properly on the light simulation software. next up Raytracing in one weekend but in a shader and shader buffer.

for more information the error was due to the way i was getting the value from the glm::mat4 of the inverse view matrix