broken framebuffer update
framebuffer don't work, no idea why, it look like nothing is being written to the texture bind to the buffer. will open a stackoverflow or a computer graphic topic about it.

still no progressive render, so putting this aside for a bit and looking for putting in place a proper UI.
#8